separated mainwindow from main.py, and added __init__.py to modules :3
